Change from "toolbox" to "class"
I'm no longer going to talk about this code as a toolbox as it's not really large/complex enough to justify the name. The toolbox distribution will remain in the github releases as it can make this coe easier to install, but otherwise we'll refer to this as a class from this time onwards.